BFM-26 Added reports and edited CSS styles
