+++ src/components/Account.vue
@@ -1,162 +1,135 @@
 <template>
-    <div id="historial">
-        <div class="container">
-            <h2>
-                Titular de la Cuenta:
-                <span>{{ username }}</span>
-            </h2>
-            <h2>
-                Saldo:
-                <span>${{ accountByUsername.balance }} COP</span>
-            </h2>
-            <h2>
-                Último Movimiento:
-                <span>
-                    {{ accountByUsername.lastChange.substring(0,10) }}
-                    {{ accountByUsername.lastChange.substring(11,19) }}
-                </span>
-            </h2>
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
+    <div class="information">
+        <div class="home">
+            <h1>Bienvenido <span>{{ employeeDetailById.name }}</span>!</h1>
         </div>
-
-        <h2>Transacciones</h2>
-        <div class="container-table">
-            <table>
-                <tr>
-                    <th>Fecha</th>
-                    <th>Hora </th>
-                    <th>Origen</th>
-                    <th>Destino</th>
-                    <th>Valor</th>
-                    <th>Nota</th>
-                    <th>Editar</th>
-                    <th>Eliminar</th>
-                </tr>
-                <tr v-for="transaction in transactionByUsername" :key="transaction.id">
-                    <td>{{ (transaction.date.substring(0,10)) }}</td>
-                    <td>{{ transaction.date.substring(11,19) }}</td>
-                    <td>{{ transaction.usernameOrigin }}</td>
-                    <td>{{ transaction.usernameDestiny }}</td>
-                    <td>${{ transaction.value }} COP</td>
-                    <td>{{ transaction.note }}</td>
-                    <td></td>
-                    <td></td>
-                </tr>
-            </table>
+        <div class="details">
+            <h3>Su información registrada es la siguiente.</h3>
+            <h2>
+                Cédula:
+                <span>{{ employeeDetailById.cedula }}</span>
+            </h2>
+            <h2>
+                Nombre de Usuario:
+                <span>{{ employeeDetailById.username }}</span>
+            </h2>
+            <h2>
+                Correo electrónico:
+                <span>{{ employeeDetailById.email }}</span>
+            </h2>
+            <h2>
+                Sexo:
+                <span>{{ employeeDetailById.gender }}</span>
+            </h2>
+                <button class="buttonUpdateEmployee" v-on:click="clickUpdateEmployee(userId)">Actualizar mi información</button>
         </div>
     </div>
 </template>
 
 
 <script>
-    import gql from 'graphql-tag';
+    import jwt_decode from "jwt-decode";
+    import gql        from "graphql-tag";
     export default{
-        name: "Account",
+        name: "Home",
+
         data: function(){
             return {
+                userId         : jwt_decode(localStorage.getItem("tokenRefresh")).user_id,
                 username: localStorage.getItem("username") || "none",
-                transactionByUsername: [],
-                accountByUsername: {
-                    balance    : 0,
-                    lastChange : ""
+                employeeDetailById : {
+                    cedula   : "",
+                    name     : "",
+                    email    : "",
+                    username : "",
+                    gender   : ""
                 }
             }
         },
         apollo: {
-            transactionByUsername: {
+            employeeDetailById : {
                 query: gql`
-                    query Query($username: String!) {
-                      transactionByUsername(username: $username) {
-                        id
-                        usernameOrigin
-                        usernameDestiny
-                        value
-                        note
-                        date
-                      }
-                    }
+                    query EmployeeDetailById($userId: Int!) {
+                        employeeDetailById(userId: $userId) {
+                            name
+                            cedula
+                            email
+                            username
+                            gender
+                        }
+                    }               
                 `,
                 variables() {
                     return {
-                        username: this.username,
+                        userId: this.userId,
                     }
                 }
             },
-            accountByUsername: {
-                query: gql`
-                    query Query($username: String!) {
-                      accountByUsername(username: $username) {
-                        balance
-                        lastChange
-                      }
-                    }
-                `,
-                variables() {
-                    return {
-                        username: this.username,
-                    }
-                }
+        }, 
+        methods: {
+            clickUpdateEmployee: async function (idm, employeeDetailById) {
+                await this.$router.push({name: "updateEmployee", params: {idm: idm}});
             }
-        },
-        created: function(){
-            this.$apollo.queries.transactionByUsername.refetch();
-            this.$apollo.queries.accountByUsername.refetch();
+
         }
-    }
+    };
 </script>
 
+
 <style>
-    #historial {
+    .information {
+        margin: 0;
+        padding: 0%;
         width: 100%;
+        height: 100%;
         display: flex;
-        justify-content: flex-start;
-        align-items: center;
         flex-direction: column;
+        justify-content: center;
+        align-items: center;
     }
-    #historial .container-table{
-        width:65S%;
-        max-height: 250px;
-        overflow-y: scroll;
-        overflow-x: hidden;
-    }
-    #historial table {
-        width: 100%;
-        border-collapse: collapse;
-        border: 1px solid rgba(0, 0, 0, 0.3);
-    }
-    #historial table td,
-    #historial table th {
-        border: 1px solid #ddd;
-        padding: 8px;
-    }
-    #historial table tr:nth-child(even) {
-        background-color: #f2f2f2;
-    }
-    #historial table tr:hover {
-        background-color: #ddd;
-    }
-    #historial table th {
-        padding-top: 12px;
-        padding-bottom: 12px;
-        text-align: left;
-        background-color: crimson;
-        color: white;
-    }
-    #historial > h2 {
-        color: #283747;
-        font-size: 25px;
-    }
-    #historial .container {
-        padding: 30px;
-        border: 3px solid rgba(0, 0, 0, 0.3);
-        border-radius: 20px;
-        margin: 5% 0 1% 0;
-    }
-    #historial .container h2 {
-        font-size: 25px;
+    
+    .information h1 {
+        font-size: 60px;
         color: #283747;
     }
-    #historial .container span {
-        color: crimson;
+    .information h2 {
+        font-size: 40px;
+        color: #283747;
+    }
+    .information span {
+        color: rgb(43, 40, 41);
         font-weight: bold;
+    }
+    .details h3 {
+        font-size: 35px;
+        color: #283747;
+        text-align: center;
+    }
+    .details h2 {
+        font-size: 35px;
+        color: #283747;
+    }
+    .details {
+        border: 0px solid rgba(0, 0, 0, 0.3);
+        border-radius: 20px;
+        padding: 30px 80px;
+        margin: 30px 0 0 0;
+    }
+    .buttonUpdateEmployee{
+        color: #000000;
+        font-size: 18px;
+        background: #e5e7e9;
+        border: 1px solid #e5e7e9;
+        border-radius: 5px;
+        padding: 10px 20px;
+        margin-right: 10px;
+    }
+    .buttonUpdateEmployee:hover{
+        color: #000000;
+        font-size: 18px;        
+        font-weight: 900;
+        background: #78a3ee;
+        border: 1px solid #e5e7e9;
     }
 </style>
